+++ zaifapi/token.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from zaifapi.api_common import get_response
 from .core import ZaifApi
-from zaifapi.url import ApiUrl
+from zaifapi.api_common import ApiUrl
 
 
 class ZaifTokenApi(ZaifApi):
